occupation controller, redesigned offer service and other changes

+++ src/main/java/com/gb/agile/craft_master/core/interfaces/CredentialService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface CredentialService {
 
-    List<Credential> getAllCredentialByUserId(Integer userId);
+    List<Credential> getAllCredentialByUserId(Long userId);
 
-    Credential addCredential(Integer userId, CredentialDto credentialDto);
+    Credential addCredential(Long userId, CredentialDto credentialDto);
 
-    void deleteCredential(Integer userId, String code);
+    void deleteCredential(Long userId, String code);
 }
